refactor(ui): extract FeatureEmptyState and scores/experiments empty states into shared components
Moves FeatureEmptyState from the dashboard into ui/packages/components
with tracking injected via callback props (onView/onPromptCopy/
onExampleCopy/onDocsLinkClick) instead of importing the dashboard's
Segment wrapper, and adds a contentOverride slot that replaces the
onboarding body while keeping the title/description/docs link.

Scores and Experiments empty-state content and views move to the shared
package so the dashboard and dev server render identical pages from one
source. The dashboard keeps thin wrappers preserving its exact props and
Segment event surface.
