+++ ui/packages/components/src/FeatureEmptyState/scoresEmptyStateContent.tsx
@@ -1,0 +1,59 @@
+import {
+  RiCheckboxCircleLine,
+  RiFilterLine,
+  RiLineChartLine,
+  RiQuestionLine,
+} from '@remixicon/react';
+
+import type { FeatureEmptyStateValueProp } from './FeatureEmptyState';
+
+export const DOCS_URL =
+  'https://www.inngest.com/docs/features/inngest-functions/steps-workflows/scoring?ref=app-empty-scores';
+
+export const DESCRIPTION =
+  'Use scores to track and evaluate custom metrics from inside your functions. Record numeric or boolean scores on any run - eval pass/fail, confidence intervals, latency, tool use. Use Inngest to measure quality and performance trends over time.';
+
+export const EXAMPLE = `import { Inngest } from 'inngest';
+import { scoreMiddleware } from 'inngest/experimental';
+
+const inngest = new Inngest({
+  id: 'my-app',
+  middleware: [scoreMiddleware()],
+});
+
+export default inngest.createFunction(
+  { id: 'grade-response' },
+  { event: 'ai/response.generated' },
+  async ({ event, step }) => {
+    const result = await step.run('agent-result', () => {/* ... */});
+
+    // Numeric and boolean scores both chart on the Scores tab
+    await step.score('quality', { name: 'response_quality', value: result.qualityScore });
+    await step.score('budget', { name: 'in_budget', value: result.tokens_used < TOKEN_BUDGET });
+  },
+);`;
+
+export const PROMPT = `Read the docs about Inngest scores @https://www.inngest.com/docs-markdown/features/inngest-functions/steps-workflows/scoring and show me how to implement scores using scoreMiddleware(), step.score(), and inngest.score() and explore what scores might be ideal for my existing functions including quality, performance, tool use.`;
+
+export const VALUE_PROPS: FeatureEmptyStateValueProp[] = [
+  {
+    icon: RiCheckboxCircleLine,
+    title: 'Grade AI & evals',
+    description: 'Record boolean or numeric scores to track quality.',
+  },
+  {
+    icon: RiLineChartLine,
+    title: 'Trend any metric',
+    description: 'Chart latency, eval results, or confidence over time.',
+  },
+  {
+    icon: RiQuestionLine,
+    title: 'Query with Insights',
+    description: 'Extract datasets by querying score data directly.',
+  },
+  {
+    icon: RiFilterLine,
+    title: 'Slice by function',
+    description: 'Filter scores per function to spot regressions fast.',
+  },
+];
